+++ lib/libesp32/Berry-0.1.10/src/be_libs.c
@@ -13,6 +13,9 @@ extern void be_load_maplib(bvm *vm);
 extern void be_load_rangelib(bvm *vm);
 extern void be_load_filelib(bvm *vm);
 
+extern void be_load_tasmota_ntvlib(bvm *vm);
+extern void be_load_wirelib(bvm *vm);
+
 void be_loadlibs(bvm *vm)
 {
     be_load_baselib(vm);
@@ -21,5 +24,8 @@ void be_loadlibs(bvm *vm)
     be_load_maplib(vm);
     be_load_rangelib(vm);
     be_load_filelib(vm);
+    be_load_byteslib(vm);
 #endif
+    be_load_tasmota_ntvlib(vm);
+    be_load_wirelib(vm);
 }