+++ lib/libesp32/Berry-0.1.10/src/port/be_tasmotalib.c
@@ -13,6 +13,7 @@ extern int l_millis(bvm *vm);
 extern int l_timereached(bvm *vm);
 extern int l_yield(bvm *vm);
 extern int l_delay(bvm *vm);
+extern int l_scaleuint(bvm *vm);
 
 extern int l_respCmnd(bvm *vm);
 extern int l_respCmndStr(bvm *vm);
@@ -22,33 +23,49 @@ extern int l_respCmndFailed(bvm *vm);
 extern int l_resolveCmnd(bvm *vm);
 
 extern int l_getlight(bvm *vm);
+extern int l_getpower(bvm *vm);
+extern int l_setlight(bvm *vm);
+extern int l_setpower(bvm *vm);
 
 // #if !BE_USE_PRECOMPILED_OBJECT
 #if 1           // TODO we will do pre-compiled later
+// Class definition
+void be_load_tasmota_ntvlib(bvm *vm)
+{
+    static const bnfuncinfo members[] = {
+        { "_op", NULL },
+        { "_opf", NULL },
+        { "_operators", NULL },
+        { "_rules", NULL },
+        { "_timers", NULL },
+        { "_cmd", NULL },
+        { "getfreeheap", l_getFreeHeap },
+        { "publish", l_publish },
+        { "cmd", l_cmd },
+        { "getoption", l_getoption },
+        { "millis", l_millis },
+        { "timereached", l_timereached },
+        { "yield", l_yield },
+        { "delay", l_delay },
+        { "scaleuint", l_scaleuint },
 
-be_native_module_attr_table(tasmota_ntv) {
-    be_native_module_function("getfreeheap", l_getFreeHeap),
-    be_native_module_function("publish", l_publish),
-    be_native_module_function("cmd", l_cmd),
-    be_native_module_function("getoption", l_getoption),
-    be_native_module_function("millis", l_millis),
-    be_native_module_function("timereached", l_timereached),
-    be_native_module_function("yield", l_yield),
-    be_native_module_function("delay", l_delay),
+        { "respcmnd", l_respCmnd },
+        { "respcmndstr", l_respCmndStr },
+        { "respcmnd_done", l_respCmndDone },
+        { "respcmnd_error", l_respCmndError },
+        { "respcmnd_failed", l_respCmndFailed },
+        { "resolvecmnd", l_resolveCmnd },
 
-    be_native_module_function("respcmnd", l_respCmnd),
-    be_native_module_function("respcmndstr", l_respCmndStr),
-    be_native_module_function("respcmnd_done", l_respCmndDone),
-    be_native_module_function("respcmnd_error", l_respCmndError),
-    be_native_module_function("respcmnd_failed", l_respCmndFailed),
-    be_native_module_function("resolvecmnd", l_resolveCmnd),
+        { "getlight", l_getlight },
+        { "getpower", l_getpower },
+        { "setlight", l_setlight },
+        { "setpower", l_setpower },
+        
+        { NULL, NULL }
+    };
+    be_regclass(vm, "Tasmota_ntv", members);
+}
 
-    be_native_module_function("getlight", l_getlight),
-
-    be_native_module_str("_operators",  "=<>!|"),
-};
-
-be_define_native_module(tasmota_ntv, NULL);
 #else
 /* @const_object_info_begin
 module tasmota (scope: global, depend: 1) {