+++ lib/libesp32/Berry-0.1.10/src/port/be_wirelib.c
@@ -7,6 +7,8 @@
  *******************************************************************/
 #include "be_object.h"
 
+extern int b_wire_init(bvm *vm);
+
 extern int b_wire_begintransmission(bvm *vm);
 extern int b_wire_endtransmission(bvm *vm);
 extern int b_wire_requestfrom(bvm *vm);
@@ -21,19 +23,25 @@ extern int b_wire_validread(bvm *vm);
 
 // #if !BE_USE_PRECOMPILED_OBJECT
 #if 1           // TODO we will do pre-compiled later
-be_native_module_attr_table(wire) {
-    be_native_module_function("_begintransmission", b_wire_begintransmission),
-    be_native_module_function("_endtransmission", b_wire_endtransmission),
-    be_native_module_function("_requestfrom", b_wire_requestfrom),
-    be_native_module_function("_available", b_wire_available),
-    be_native_module_function("_write", b_wire_write),
-    be_native_module_function("_read", b_wire_read),
-    be_native_module_function("scan", b_wire_scan),
-    be_native_module_function("write", b_wire_validwrite),
-    be_native_module_function("read", b_wire_validread),
-};
-
-be_define_native_module(wire, NULL);
+void be_load_wirelib(bvm *vm)
+{
+    static const bnfuncinfo members[] = {
+        { "_bus", NULL },               // bus number
+        { "init", b_wire_init },
+        { "_begintransmission", b_wire_begintransmission },
+        { "_endtransmission", b_wire_endtransmission },
+        { "_requestfrom", b_wire_requestfrom },
+        { "_available", b_wire_available },
+        { "_write", b_wire_write },
+        { "_read", b_wire_read },
+        { "scan", b_wire_scan },
+        { "write", b_wire_validwrite },
+        { "read", b_wire_validread },
+        
+        { NULL, NULL }
+    };
+    be_regclass(vm, "Wire", members);
+}
 #else
 /* @const_object_info_begin
 module tasmota (scope: global, depend: 1) {